+++ src/peripherals/DACPeripheral.hpp
@@ -10,7 +10,7 @@ extern "C" {
 class DACPeripheral : public Peripheral<DAC_TypeDef> {
 public:
   void configure() override;
-  void start(uint16_t *dacBuffer, uint16_t dataLength);
+  void start(uint16_t *dacBuffer, int dataLength);
   DAC_TypeDef *getPeripheral() const override;
 
 private:
